Move back to one e2e suite
