+++ .github/workflows/ci.yaml
@@ -139,7 +139,7 @@ jobs:
       max-parallel: 7
       matrix:
         driver: [jetstream]
-        case: [e2e-suite-1, e2e-suite-2, kafka-e2e, http-e2e, nats-e2e, sdks-e2e, reduce-e2e]
+        case: [e2e, e2e, kafka-e2e, http-e2e, nats-e2e, sdks-e2e, reduce-e2e]
     steps:
       - name: Checkout code
         uses: actions/checkout@v3
