+++ Makefile
@@ -102,8 +102,8 @@ test-coverage-with-isb:
 test-code:
 	go test -tags=isb_redis -race -v $(shell go list ./... | grep -v /vendor/ | grep -v /numaflow/test/)
 
-test-e2e-suite-1:
-test-e2e-suite-2:
+test-e2e:
+test-e2e:
 test-kafka-e2e:
 test-http-e2e:
 test-nats-e2e:
@@ -136,7 +136,7 @@ Test%:
 	kubectl -n numaflow-system delete po -lapp.kubernetes.io/component=controller-manager,app.kubernetes.io/part-of=numaflow
 	kubectl -n numaflow-system delete po e2e-api-pod  --ignore-not-found=true
 	cat test/manifests/e2e-api-pod.yaml |  sed 's@quay.io/numaproj/@$(IMAGE_NAMESPACE)/@' | sed 's/:$(BASE_VERSION)/:$(VERSION)/' | kubectl -n numaflow-system apply -f -
-	-go test -v -timeout 10m -count 1 --tags test -p 1 ./test/e2e-suite-1  -run='.*/$*'
+	-go test -v -timeout 10m -count 1 --tags test -p 1 ./test/e2e  -run='.*/$*'
 	$(MAKE) cleanup-e2e
 
 .PHONY: ui-build
