+++ test/e2e/testdata/simple-pipeline.yaml
