+++ test/e2e/testdata/watermark.yaml
